+++ tests/test_checker/test_noqa.py
@@ -245,6 +245,7 @@ SHOULD_BE_RAISED = types.MappingProxyType(
         'WPS478': 1,
         'WPS479': 0,
         'WPS480': 0,  # only triggers on 3.12+
+        'WPS481': 3,
         'WPS500': 1,
         'WPS501': 1,
         'WPS502': 0,  # disabled since 1.0.0